:recycle: Split user-agent extra tests
Move the optional-extra failure coverage for the user_agent context processor and UserAgentMixin into focused test modules.

Keep the existing broader context processor and view mixin tests focused on their original behavior while preserving the same assertions for missing user-agents.
