+++ tests/tests/context_processors/tests.py
@@ -1,8 +1,5 @@
 import os
-from unittest.mock import patch
 
-from django.core.exceptions import ImproperlyConfigured
-from django.test import RequestFactory
 from django.test import override_settings
 
 from django_boost.test import TestCase
@@ -71,21 +68,3 @@ class TestConTextProcessors(TestCase):
                     self.assertContains(response, value)
                 else:
                     self.assertNotContains(response, value)
-
-    def test_user_agent_requires_extra(self):
-        from django_boost.context_processors import user_agent
-
-        request = RequestFactory().get('/')
-        real_import = __import__
-
-        def import_without_user_agents(name, *args, **kwargs):
-            if name == 'user_agents':
-                raise ImportError(name)
-            return real_import(name, *args, **kwargs)
-
-        with patch('builtins.__import__', import_without_user_agents):
-            with self.assertRaisesMessage(
-                ImproperlyConfigured,
-                "pip install django-boost[useragent]",
-            ):
-                user_agent(request)
